+++ src/signalling/room_service.rs
@@ -577,7 +577,7 @@ impl Handler<DeleteElements<Validated>> for RoomService {
 
         if !room_messages_futs.is_empty() {
             future::try_join_all(room_messages_futs)
-                .map_ok(|_| ())
+                .map_ok(drop)
                 .map_err(RoomServiceError::RoomMailboxErr)
                 .boxed_local()
         } else if !deletes_from_room.is_empty() {
@@ -587,7 +587,7 @@ impl Handler<DeleteElements<Validated>> for RoomService {
                 || future::ok(()).boxed_local(),
                 |room| {
                     room.send(Delete(deletes_from_room))
-                        .map_ok(|_| ())
+                        .map_ok(drop)
                         .map_err(RoomServiceError::RoomMailboxErr)
                         .err_into()
                         .boxed_local()