+++ tests/e2e/object/jason.rs
@@ -30,8 +30,29 @@ impl Object<Jason> {
             r#"
                 async (jason) => {
                     let room = await jason.init_room();
-                    room.on_failed_local_media(() => {});
-                    room.on_connection_loss(() => {});
+                    let onFailedLocalStreamListener = {
+                        subs: [],
+                        count: 0
+                    };
+                    room.on_failed_local_media(() => {
+                        onFailedLocalStreamListener.count++;
+                        onFailedLocalStreamListener.subs =
+                            onFailedLocalStreamListener.subs
+                                .filter((sub) => sub());
+                    });
+                    let connLossListener = {
+                        isLost: false,
+                        subs: []
+                    };
+                    room.on_connection_loss(async (recon) => {
+                        connLossListener.isLost = true;
+                        for (sub of connLossListener.subs) {
+                            sub();
+                        }
+                        connLossListener.subs = [];
+                        await recon.reconnect_with_backoff(100, 1.0, 100);
+                        connLossListener.isLost = false;
+                    });
                     let closeListener = {
                         closeReason: null,
                         isClosed: false,
@@ -66,7 +87,9 @@ impl Object<Jason> {
                     return {
                         room: room,
                         closeListener: closeListener,
-                        localTracksStore: localTracksStore
+                        localTracksStore: localTracksStore,
+                        connLossListener: connLossListener,
+                        onFailedLocalStreamListener: onFailedLocalStreamListener
                     };
                 }
             "#,
@@ -89,7 +112,7 @@ impl Object<Jason> {
             [room.ptr()],
         ))
         .await
-        .map(|_| ())
+        .map(drop)
     }
 
     /// Drops [`Jason`] API object, so all the related objects (rooms,
@@ -105,6 +128,6 @@ impl Object<Jason> {
             [],
         ))
         .await
-        .map(|_| ())
+        .map(drop)
     }
 }