+++ tests/e2e/steps/room.rs
@@ -1,4 +1,4 @@
-use cucumber_rust::{then, when};
+use cucumber_rust::{given, then, when};
 
 use crate::world::World;
 
@@ -27,4 +27,49 @@ async fn when_room_closed_by_client(world: &mut World, id: String) {
 #[when(regex = r"^(\S+) disposes Jason object$")]
 async fn when_jason_object_disposes(world: &mut World, id: String) {
     world.dispose_jason(&id).await.unwrap();
+}
+
+#[given(regex = r"^(\S+)'s `getUserMedia\(\)` (audio |video )?errors$")]
+async fn given_member_gum_will_error(
+    world: &mut World,
+    id: String,
+    kind: String,
+) {
+    let member = world.get_member(&id).unwrap();
+    let media_devices = member.media_devices_mock();
+    let (video, audio) = if kind.is_empty() {
+        (true, true)
+    } else {
+        (kind.contains("video"), kind.contains("audio"))
+    };
+    media_devices.mock_gum(video, audio).await;
+}
+
+#[when(regex = "^(\\S+) enables (video|audio|video and audio) in local \
+                 media settings$")]
+async fn when_member_enables_via_local_media_settings(
+    world: &mut World,
+    id: String,
+    kind: String,
+) {
+    let member = world.get_member(&id).unwrap();
+    let video = kind.contains("video");
+    let audio = kind.contains("audio");
+    let _ = member
+        .room()
+        .set_local_media_settings(video, audio)
+        .await
+        .unwrap();
+}
+
+#[then(regex = "^(\\S+)'s `Room.on_failed_local_stream\\(\\)` fires (\\d+) \
+                 time(:?s)?$")]
+async fn then_room_failed_local_stream_fires(
+    world: &mut World,
+    id: String,
+    times: u64,
+) {
+    let member = world.get_member(&id).unwrap();
+    member.room().when_failed_local_stream_count(times).await;
+    assert!(true);
 }