+++ tests/e2e/steps/track.rs
@@ -207,3 +207,10 @@ async fn then_member_doesnt_have_remote_tracks_with(
     let tracks_count = tracks_store.count().await.unwrap();
     assert_eq!(tracks_count, 0);
 }
+
+#[then(regex = r"^(\S+) doesn't have live tracks$")]
+async fn then_member_doesnt_has_live_tracks(world: &mut World, id: String) {
+    let member = world.get_member(&id).unwrap();
+    let local_tracks = member.room().local_tracks().await.unwrap();
+    assert!(local_tracks.is_all_tracks_ended().await.unwrap());
+}