+++ tests/e2e/world/mod.rs
@@ -195,11 +195,10 @@ impl World {
                 self.control_client.create(&path, element).await?;
             }
         }
-        let jason =
-            Object::spawn(Jason, self.window_factory.new_window().await)
-                .await?;
+        let window = self.window_factory.new_window().await;
+        let jason = Object::spawn(Jason, window.clone()).await?;
         let room = jason.init_room().await?;
-        let member = builder.build(room).await?;
+        let member = builder.build(room, window).await?;
 
         self.jasons.insert(member.id().to_owned(), jason);
         self.members.insert(member.id().to_owned(), member);